Move css files to their folder
